refactor: linter fixes
